refactor($User): DTO, Entity 분리
