+++ src/main/java/gift/dto/UserRequestDto.java
@@ -1,0 +1,11 @@
+package gift.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+
+public record UserRequestDto(
+    @Email String email,
+    @NotEmpty String password
+) {
+
+}
