+++ src/main/java/gift/dto/UserResponseDto.java
@@ -1,0 +1,11 @@
+package gift.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+
+public record UserResponseDto(
+    @NotNull Long id,
+    @Email String email
+) {
+
+}
