+++ src/main/java/gift/entity/User.java
@@ -1,11 +1,12 @@
-package gift.model;
+package gift.entity;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 
-public record User (
+public record User(
+    @NotNull Long id,
     @Email String email,
     @NotEmpty String password
 ) {
-
 }
